Partial breakdown into fragments (Not working currently)
