+++ app/src/main/java/dev/digitaldreamweavers/qrguard/ui/camera/FloatingScanButtonViewModel.java
@@ -1,0 +1,7 @@
+package dev.digitaldreamweavers.qrguard.ui.camera;
+
+import androidx.lifecycle.ViewModel;
+
+public class FloatingScanButtonViewModel extends ViewModel {
+    // TODO: Implement the ViewModel
+}
