+++ app/src/main/res/layout/activity_main.xml
@@ -18,23 +18,11 @@
         <FrameLayout
             android:layout_width="match_parent"
             android:layout_height="0dp"
-            android:layout_weight="1">
-            <androidx.camera.view.PreviewView
-                android:id="@+id/viewFinder"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent">
-            </androidx.camera.view.PreviewView>
-            <com.google.android.material.floatingactionbutton.ExtendedFloatingActionButton
-                android:id="@+id/scanButton"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="bottom|center_horizontal"
-                android:clickable="false"
-                style="?attr/extendedFloatingActionButtonSurfaceStyle"
-                android:text="@string/qr_status_waiting"
-                android:contentDescription="@string/perform_a_scan_on_this_qr_code"
-                android:layout_marginBottom="16dp" />
+            android:layout_weight="1"
+            android:id="@+id/frame_container">
+
         </FrameLayout>
+
         <com.google.android.material.bottomnavigation.BottomNavigationView
             android:id="@+id/bottom_navigation"
             android:layout_width="match_parent"
